streamlit_page/requirements.txt
altair==5.5.0 ; python_version >= "3.13"
attrs==25.3.0 ; python_version >= "3.13"
blinker==1.9.0 ; python_version >= "3.13"
cachetools==6.1.0 ; python_version >= "3.13"
certifi==2025.6.15 ; python_version >= "3.13"
charset-normalizer==3.4.2 ; python_version >= "3.13"
click==8.2.1 ; python_version >= "3.13"
colorama==0.4.6 ; python_version >= "3.13" and platform_system == "Windows"
contourpy==1.3.2 ; python_version >= "3.13"
cycler==0.12.1 ; python_version >= "3.13"
fonttools==4.58.4 ; python_version >= "3.13"
gitdb==4.0.12 ; python_version >= "3.13"
gitpython==3.1.44 ; python_version >= "3.13"
idna==3.10 ; python_version >= "3.13"
jinja2==3.1.6 ; python_version >= "3.13"
jsonschema-specifications==2025.4.1 ; python_version >= "3.13"
jsonschema==4.24.0 ; python_version >= "3.13"
kiwisolver==1.4.8 ; python_version >= "3.13"
markupsafe==3.0.2 ; python_version >= "3.13"
matplotlib==3.10.3 ; python_version >= "3.13"
narwhals==1.45.0 ; python_version >= "3.13"
numpy==2.3.1 ; python_version >= "3.13"
packaging==25.0 ; python_version >= "3.13"
pandas==2.3.0 ; python_version >= "3.13"
pillow==11.3.0 ; python_version >= "3.13"
plotly==6.2.0 ; python_version >= "3.13"
protobuf==6.31.1 ; python_version >= "3.13"
psycopg2-binary==2.9.10 ; python_version >= "3.13"
pyarrow==20.0.0 ; python_version >= "3.13"
pydeck==0.9.1 ; python_version >= "3.13"
pyparsing==3.2.3 ; python_version >= "3.13"
python-dateutil==2.9.0.post0 ; python_version >= "3.13"
python-dotenv==1.1.1 ; python_version >= "3.13"
pytz==2025.2 ; python_version >= "3.13"
referencing==0.36.2 ; python_version >= "3.13"
requests==2.32.4 ; python_version >= "3.13"
rpds-py==0.26.0 ; python_version >= "3.13"
six==1.17.0 ; python_version >= "3.13"
smmap==5.0.2 ; python_version >= "3.13"
streamlit==1.46.1 ; python_version >= "3.13"
tenacity==9.1.2 ; python_version >= "3.13"
toml==0.10.2 ; python_version >= "3.13"
tornado==6.5.1 ; python_version >= "3.13"
typing-extensions==4.14.0 ; python_version >= "3.13"
tzdata==2025.2 ; python_version >= "3.13"
urllib3==2.5.0 ; python_version >= "3.13"
watchdog==6.0.0 ; platform_system != "Darwin" and python_version >= "3.13"
